# working-scripts/outputs/api-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MVE Resource Server API",
    "version": "1.0.0",
    "description": "MVE (Meta Virtual Environment) Resource Server API\n\n이 API는 다음 기능을 제공합니다:\n- 🎵 음원 파일 관리 (스트리밍, 업로드, 검색)\n- 🎨 3D 모델 파일 관리 (GLB)\n- 🤖 AI 기반 3D 모델 생성\n- 🎭 아바타 악세서리 프리셋 관리\n- 🎪 콘서트 세션 관리\n\n**인증**: 모든 API는 JWT Bearer 토큰 인증이 필요합니다.\n**토큰 발급**: [MVE Login Server](https://github.com/prayslaks/mve-login-server)에서 로그인하여 JWT 토큰을 발급받으세요.",
    "contact": {
      "name": "MVE Development Team",
      "url": "https://github.com/prayslaks"
    },
    "license": {
      "name": "ISC",
      "url": "https://opensource.org/licenses/ISC"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "로컬 개발 서버"
    },
    {
      "url": "http://your-ec2-public-ip",
      "description": "AWS EC2 프로덕션 서버 (HTTP)"
    },
    {
      "url": "https://your-domain.com",
      "description": "AWS EC2 프로덕션 서버 (HTTPS)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "MVE Login Server에서 발급받은 JWT 토큰을 입력하세요."
      }
    },
    "schemas": {
...
      "AudioFile": {
        "type": "object",
        "description": "음원 파일 정보",
        "properties": {
          "id": {
            "type": "integer",
            "description": "음원 ID",
            "example": 1
          },
          "title": {
            "type": "string",
            "description": "음원 제목",
            "example": "Sample Track 1"
          },
          "artist": {
            "type": "string",
            "nullable": true,
            "description": "아티스트",
            "example": "Artist A"
          },
          "file_path": {
            "type": "string",
            "description": "파일 경로",
            "example": "audio/sample1.m4a"
          },
          "file_size": {
            "type": "integer",
            "description": "파일 크기 (bytes)",
            "example": 3145728
          },
          "duration": {
            "type": "integer",
            "nullable": true,
            "description": "재생 시간 (초)",
            "example": 180
          },
          "format": {
            "type": "string",
            "description": "파일 포맷",
            "example": "m4a"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "생성 시간"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/audio/list": {
      "get": {
        "summary": "음원 목록 조회",
        "description": "모든 음원 파일 목록을 조회합니다 (공용 리소스, JWT 인증 필요)",
        "tags": [
          "Audio"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "음원 목록 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "count": {
                      "type": "integer",
                      "example": 3
                    },
                    "audio_files": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AudioFile"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 실패"
          },
          "500": {
            "description": "서버 오류"
          }
        }
      }
    },
    "/api/audio/stream/{id}": {
      "get": {
        "summary": "음원 스트리밍 URL 획득",
        "description": "음원 스트리밍 URL을 반환합니다 (S3 환경에서는 Presigned URL, 로컬 환경에서는 내부 URL 제공)",
        "tags": [
          "Audio"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "음원 ID",
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "스트리밍 URL 생성 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "stream_url": {
                      "type": "string",
                      "description": "스트리밍 URL (S3 Presigned URL 또는 로컬 URL)",
                      "example": "https://bucket.s3.amazonaws.com/audio/file.aac?X-Amz-Signature=..."
                    },
                    "audio_file": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "title": {
                          "type": "string"
                        },
                        "format": {
                          "type": "string"
                        },
                        "file_size": {
                          "type": "integer"
                        }
                      }
                    },
                    "expires_in": {
                      "type": "integer",
                      "description": "URL 유효 시간 (초, S3만 해당)",
                      "example": 3600
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 실패"
          },
          "404": {
            "description": "음원을 찾을 수 없음"
          },
          "500": {
            "description": "서버 오류"
          }
        }
      }
    },
    "/api/audio/file/{id}": {
      "get": {
        "summary": "로컬 음원 파일 직접 스트리밍",
        "description": "로컬 스토리지에서 음원 파일을 직접 스트리밍합니다 (Range Request 지원)",
        "tags": [
          "Audio"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "음원 ID",
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "전체 파일 스트리밍",
            "content": {
              "audio/*": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "206": {
            "description": "부분 콘텐츠 (Range Request)",
            "headers": {
              "Content-Range": {
                "schema": {
                  "type": "string"
                },
                "description": "바이트 범위"
              }
            }
          },
          "401": {
            "description": "인증 실패"
          },
          "404": {
            "description": "음원을 찾을 수 없음"
          }
        }
      }
    },
    "/api/audio/upload": {
      "post": {
        "summary": "음원 파일 업로드",
        "description": "음원 파일을 업로드합니다 (AAC, M4A, MP3, WAV 지원, 최대 100MB)",
        "tags": [
          "Audio"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "audio",
                  "title"
                ],
                "properties": {
                  "audio": {
                    "type": "string",
                    "format": "binary",
                    "description": "음원 파일 (AAC, M4A, MP3, WAV)"
                  },
                  "title": {
                    "type": "string",
                    "description": "음원 제목",
                    "example": "Sample Track"
                  },
                  "artist": {
                    "type": "string",
                    "description": "아티스트 이름 (선택)",
                    "example": "Artist Name"
                  },
                  "duration": {
                    "type": "number",
                    "description": "재생 시간 (초, 선택)",
                    "example": 180
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "음원 업로드 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Audio file uploaded successfully"
                    },
                    "audio_file": {
                      "$ref": "#/components/schemas/AudioFile"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청 (파일 또는 제목 누락)"
          },
          "401": {
            "description": "인증 실패"
          },
          "500": {
            "description": "서버 오류"
          }
        }
      }
    },
    "/api/audio/search/{query}": {
      "get": {
        "summary": "음원 검색",
        "description": "제목 또는 아티스트로 음원을 검색합니다",
        "tags": [
          "Audio"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "검색 키워드 (제목 또는 아티스트)",
            "example": "sample"
          }
        ],
        "responses": {
          "200": {
            "description": "검색 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "count": {
                      "type": "integer",
                      "example": 2
                    },
                    "audio_files": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AudioFile"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 실패"
          },
          "500": {
            "description": "서버 오류"
          }
        }
      }
    },
    "/api/audio/{id}": {
      "get": {
        "summary": "특정 음원 정보 조회",
        "description": "음원 ID로 상세 정보를 조회합니다",
        "tags": [
          "Audio"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "음원 ID",
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "음원 정보 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "audio_file": {
                      "$ref": "#/components/schemas/AudioFile"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 실패"
          },
          "404": {
            "description": "음원을 찾을 수 없음"
          },
          "500": {
            "description": "서버 오류"
          }
        }
      }
    },
...
  }
}